% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahull2lines.R
\name{ahull2lines}
\alias{ahull2lines}
\title{ahull2lines}
\usage{
ahull2lines(hull)
}
\arguments{
\item{hull}{pass through for hull object}
}
\value{
a SpatialLines object
}
\description{
gratefully pilfered from alphahull
}
\details{
Code gratefully pilfered from alphahull package so that I could expose parameters
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[Lines][sp::Lines], [SpatialLines][sp::SpatialLines]
}
