% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc2line.R
\name{arc2line}
\alias{arc2line}
\title{arc2line}
\usage{
arc2line(center, r, vector, theta, npoints = 100)
}
\arguments{
\item{center}{pass through for hull object}

\item{r}{pass through for hull object}

\item{vector}{pass through for hull object}

\item{theta}{pass through for hull object}

\item{npoints}{pass through for hull object, Default: 100}
}
\value{
sp lines
}
\description{
gratefully pilfered from alphahull
}
\details{
Code gratefully pilfered from alphahull package so that I could expose parameters
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[anglesArc][alphahull::anglesArc]
 [Line][sp::Line]
}
