% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello()
}
\value{
a print statement
}
\description{
I'm still learning, it's nice to have a little positive encouragement when this actually works.
}
\details{
The generic "print hello world" function to test install and package dev
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 holyhull::hello()
 }
}
}
