% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holyhull.R
\name{holyhull}
\alias{holyhull}
\title{FUNCTION_TITLE}
\usage{
holyhull(
  sf_frame = NULL,
  method = "test",
  alpha_value = 0.01,
  concavity = 2,
  length_threshold = 0
)
}
\arguments{
\item{sf_frame}{PARAM_DESCRIPTION, Default: NULL}

\item{method}{PARAM_DESCRIPTION, Default: 'test'}

\item{alpha_value}{PARAM_DESCRIPTION, Default: 0.01}

\item{concavity}{PARAM_DESCRIPTION, Default: 2}

\item{length_threshold}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[Lines][sp::Lines], [SpatialLines][sp::SpatialLines], [Line][sp::Line], [SpatialPolygons][sp::SpatialPolygons], [Polygons][sp::Polygons], [Polygon][sp::Polygon]
 [anglesArc][alphahull::anglesArc], [ahull][alphahull::ahull]
 [graph_from_edgelist][igraph::graph_from_edgelist], [is.connected][igraph::is.connected], [degree][igraph::degree], [clusters][igraph::clusters], [E][igraph::E], [get.shortest.paths][igraph::get.shortest.paths], [V][igraph::V]
 [st_as_sf][sf::st_as_sf], [st_set_crs][sf::st_set_crs], [st_crs][sf::st_crs], [st_convex_hull][sf::st_convex_hull], [st_union][sf::st_union], [st_transform][sf::st_transform], [st_coordinates][sf::st_coordinates]
 [concaveman][concaveman::concaveman]
}
