% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_batman_pts.R
\name{make_batman_pts}
\alias{make_batman_pts}
\title{FUNCTION_TITLE}
\usage{
make_batman_pts()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[geojson_sf][geojsonsf::geojson_sf]
}
