% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spLines2poly.R
\name{spLines2poly}
\alias{spLines2poly}
\title{spLines2poly}
\usage{
spLines2poly(sp_lines)
}
\arguments{
\item{sp_lines}{pass through for hull object}
}
\value{
sp_polys
}
\description{
gratefully pilfered and wrapper-d around alphahull
}
\details{
Code gratefully pilfered from alphahull package so that I could expose parameters
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[SpatialPolygons][sp::SpatialPolygons], [Polygons][sp::Polygons], [Polygon][sp::Polygon]
}
