% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahull2lines.R
\name{ahull2lines}
\alias{ahull2lines}
\title{FUNCTION_TITLE}
\usage{
ahull2lines(hull)
}
\arguments{
\item{hull}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[Lines][sp::Lines], [SpatialLines][sp::SpatialLines]
}
