% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc2line.R
\name{arc2line}
\alias{arc2line}
\title{FUNCTION_TITLE}
\usage{
arc2line(center, r, vector, theta, npoints = 100)
}
\arguments{
\item{center}{PARAM_DESCRIPTION}

\item{r}{PARAM_DESCRIPTION}

\item{vector}{PARAM_DESCRIPTION}

\item{theta}{PARAM_DESCRIPTION}

\item{npoints}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[anglesArc][alphahull::anglesArc]
 [Line][sp::Line]
}
