% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashape2poly.R
\name{ashape2poly}
\alias{ashape2poly}
\title{FUNCTION_TITLE}
\usage{
ashape2poly(ashape)
}
\arguments{
\item{ashape}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[graph_from_edgelist][igraph::graph_from_edgelist], [is.connected][igraph::is.connected], [degree][igraph::degree], [clusters][igraph::clusters], [E][igraph::E], [get.shortest.paths][igraph::get.shortest.paths], [V][igraph::V]
}
