% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_batman_pts.R
\name{make_batman_pts}
\alias{make_batman_pts}
\title{make_batman_pts}
\usage{
make_batman_pts()
}
\value{
geojson formatted point type FeatureCollection
}
\description{
Returns a toy set of batman themed points which are used in test plotting.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[geojson_sf][geojsonsf::geojson_sf]
}
