% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spLines2poly.R
\name{spLines2poly}
\alias{spLines2poly}
\title{FUNCTION_TITLE}
\usage{
spLines2poly(sp_lines)
}
\arguments{
\item{sp_lines}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[SpatialPolygons][sp::SpatialPolygons], [Polygons][sp::Polygons], [Polygon][sp::Polygon]
}
