% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_alpha_hull.R
\name{st_alpha_hull}
\alias{st_alpha_hull}
\title{FUNCTION_TITLE}
\usage{
st_alpha_hull(x_in_4326, y_in_4326, alpha_value = 0.01)
}
\arguments{
\item{x_in_4326}{PARAM_DESCRIPTION}

\item{y_in_4326}{PARAM_DESCRIPTION}

\item{alpha_value}{PARAM_DESCRIPTION, Default: 0.01}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[ahull][alphahull::ahull]
 [st_as_sf][sf::st_as_sf], [st_set_crs][sf::st_set_crs], [st_crs][sf::st_crs], [st_transform][sf::st_transform]
}
